Opt-in Equatable State
This allows users to opt-in to state (or any substate) being Equatable.
This causes `newState` to not be called if the selected state is `Equatable` & unchanged.

I chose to go opt-in since it has the least impact on existing users. If we wanted to make it requred, instead, It is a relatively easy change from here, by making `StateType: Equatable`, and `SelectedState: Equatable` for all `subscribe` functions.

This change is mostly backwards compatible, with no code changes necessary for existing users. The only difference is users who already have been selecting an `Equatable` sub-state will find `newState` not being called as often.

There is a performance implication here since comparisons are made each time the store changes, which can be reduced by using subscription selectors to select the minimum necessary state. (which reduces the equality checks required.)

Fixes #176