Use EquatableState protocol instead of Equatable
Instead of having automatic state checks based on `Equatable`, this now requires the user to have their selected state conform to `EquatableState` in order to get the features indicated in the `Opt-in Equatable State` commit.

A default implementation is provided when the types conform to `Equatable`.

The default implementation can be overridden by implementing `func isEqualState(to state: Self) -> Bool` on one’s selected state type. This allows one to separate the behaviours if `Equatable` doesn’t match the desired diff check.