[Subscriptions] Add SkipRepeats and Improve Separation
Subscriptions now act as a pipe between the store and subscribers (very much as signals in reactive programming libraries). The store puts all new state values into that pipe; subscribers can transform it such that certain state updates are skipped or only a part of the store’s state is sent in these updates.