Add script with the steps to integrate SwiftLint
The main reason for this is that there's no way to only fetch pods
without touching the project file. For our purposes though the project
file should not be touched, we are only interested in the SwiftLint
binary being fetched. So what we can do is just undo all the changes
made to the project by CocoaPods.