Update README example to newer reducer format
In #177, `Reducer` was changed from a protocol, requiring the user to create a struct to implement it, to a typealias, allowing one to simply write a function conforming to it. This updates the Readme example to use the new typealias based reducer instead of a struct conforming to the `Reducer` protocol.

[ci skip]