Generic ActionCreator
This splits ActionCreator out of `StoreType` and makes it a stand-alone generic typealias.
This allows one to define an action creator without the context of the store its being called on, and lets you use that action creator on any conforming store.
This allows more isolated action creators that need only know about some sub-state.

See `testAcceptsConformingActionCreators` for an example.