Pin SwiftLint using CocoaPods (#218)
* Centralize SwiftLint logic in Build Phase script

* Add a Podfile to fetch SwiftLint

* Add script with the steps to integrate SwiftLint

The main reason for this is that there's no way to only fetch pods
without touching the project file. For our purposes though the project
file should not be touched, we are only interested in the SwiftLint
binary being fetched. So what we can do is just undo all the changes
made to the project by CocoaPods.

* Move script to install SwiftLint in scripts folder

I didn't notice there was one before looking at GitHub. Also changed the
script name to match the others in the folder.

* Use repo naming convention for script to run SwiftLint

* Use dummy target for SwiftLint integration

This approach is more robust that having to rely on developers to
remembre to run the installation script, and at the same time doesn't
pollute the actual ReSwift targets with unnecessary CocoaPods build
phases and imports.

* Track xcworkspace

See https://github.com/ReSwift/ReSwift/pull/218#issuecomment-287205464

* Allow spaces in sourceroot directory

Quotes the commands for running swiftlint to allow `$SRCROOT` to contain directories with spaces.

* Update swiftlint warning to refer to cocoapods

Instead of telling user to run install-swiftlint which no longer exists.

* Update swiftlint to latest

We also don't need to pin the version in the `Podfile`, since `Podfile.lock` pins for us on `pod install`. This allows the upgrade process to be just `pod update`, instead of having to manually update the version in the Podfile.

* Add note on SwiftLint + CocoaPods setup in README