[Release] Stop Producing Binaries
Using precompiled binary frameworks is not supported by Apple. Further,
the pre-built binaries are being built with the version of Swift that is
current at the time of the release. This leads to problems of Swift version
mismatches when trying to install ReSwift via Carthage.
I suggest we stop producing and publishing binaries.