Add Subscription `skip(when:)` and `only(when:)`
These subscription filtering methods are provided as convenience methods for `skipRepeats`.

1. `skip(when:)` is a direct forward to `skipRepeats` and is simply provided as a more swift-like naming option.
2. `only(when:)` is the inverse to `skipRepeats`, and is provided for specifying when the user wants the subscription to notify, as opposed to when to skip notifying.

Both of these functions are technically "sugar", however I feel like they add value to the library by making the available use cases of `skipRepeats` more obvious to users.

For example, a user could chain calls to these, such as `$0.select({ $0.child }).skip(when: { $0.identifier == $1.identifier }).only(when: { $1.name.lowercased.hasPrefix("a") })` which would notify the subscriber when the child changes, and the new child's `name` starts with "a"