Add parameter naming & descriptions to skipRepeats
Adds internal parameter names to the closure passed to skipRepeats, and documentation for this as well.

While this does not affect autocomplete, it does provide better documention via xcode's option+click functionality, and should help users who might be confused at which state is the new / old one as the parameters are otherwise identical.