Add test example for double equatable check
Having a default implementation for `skipRepeats` when the selected state is equatable causes what could be a performance issue when the user implements their own `skipRepeats`.
The original default implementation is still called, even though a default exists.
This could also cause unwanted behaviour if the custom skipRepeats wants to compare something not checked in the default equatable.