Fix retain cycle in SubscriptionBox (#278)
* Fix retain cycle in SubscriptionBox

`SubscriptionBox` retains `originalSubscription`, which was retaining `self` in the observe closure.

We can safely make this `[unowned self]` since `SubscriptionBox` is the only object that holds a reference to `originalSubscription` (it is discarded in the `Store` function that creates the box)

* add SubscriptionBox factory as testing seam

* add test to verify the subscription is retained

* add closure retain cycle test for transformation

* Fix second retain cycle in SubscriptionBox

`SubscriptionBox` retains `originalSubscription`, which retains the transformed subscription, which was retaining `self` in the observe block, completing the cycle back to `SubscriptionBox`

We can safely make this `[unowned self]` since `SubscriptionBox` is the only object that holds a reference to `originalSubscription` (it is discarded in the `Store` function that creates the box).

* Update Travis CI image to Xcode 8.3

Resolves a compiler segfault in `StoreSubscriptionTests.swift`

* Update changelog