Proof of concept inverting subscription transforms
RE: https://github.com/ReSwift/ReSwift/issues/263

This is a proof of concept for simplifying the subscription selection / skipping usage.

```swift
// Previously this:
store.subscribe(subscriber) {
    $0.select { $0.testValue }.skip { $0 == $1 }
}
// Now can be this:
store.select { $0.testValue }
    .skip { $0 == 1 }
    .subscribe(subscriber)
```

I've implemented the prototype in such a way that maintains backwards compatibility, which would allow us to deprecate the changes for a release or two before breaking compatibility. (if we ever wanted to break it to clean up the code)