Attempt to allow splitting of the transform chain
Trying to take @DivineDominion's suggestions and allow the subscription's transform chain to be split, and only appending the subscription at the end.

Problems:
* This is getting very confusing, retains-wise. Its hard to keep track of whats retaining what, and if there are any cycles (Which was a flaw of the first system I was trying to avoid)
* Initial newState notify: Since there is an assumption we can now hold onto subscription objects, and subscribe later, we can no longer pass along `initialState` into the subscription, so we need to somehow store the transformations separate from the `notify` variable, since we only have `newState` instead of `oldState, newState` for the first notify.
* Now we have to pass along the "original subscription" since we lose it from the `subscription()` function on Store

Perhaps there is a solution involving growing the distance between `S.State` and `SelectedState` instead of chaining `PendingSubscription` objects?

Or maybe the way of solving the above problems and maintaining the "freedom of use" of the chaining that @DivineDominion describes is to not start the chain from the store, but instead to use something like:

```
Subscription().select().subscribe(self, to: store)
```

But I feel like that breaks the whole "the store is the source of everything" flow that ReSwift has going on. (and might have generics issues, anyways)

So the questions remain on my original implementation
1. How do we allow people to split off 2x somewhere in the subscription chain (Maybe we can add to `notify` instead of overriding it?)
2. How do we allow multiple `subscribe()` calls? (maybe we can just make `subscriber` an array?