Improve performance of adding subscribers
Fixes #324

When adding many subscribers, enumerating the subscribers array to remove duplicates is extremely costly.
Since we restrict `AnySubscriber` to `class` types, we can store the subscriber's `ObjectIdentifier`, and use that for a hash value & equality check, to allow us to use a set for subscribers instead of an array.

In the included performance test, this results in a 94% improvement.
However, this also regresses the state update performance by 46% due to the increased cost of filtering & enumerating a set vs an array.

Note: This will change the order of subscribers called when state changes, however the public API does not indicate that this was ordered to begin with, so I doubt it will cause issues.