Fix disappearing subscription bug from #325 (#330)
* Add performance baseline for new computer

Next commits are changes that could possibly affect these performance metrics. Updating the baselines before making the changes ensure that the changes have no noticable affect on performance.

* Add failing test for subscription mutation bug

Intentionally failing test to demonstrate a bug introduced by #325

Adding subscribers in `newState` will sometimes cause the new subscriber to be lost.

* Fix disappearing subscription bug from #325

If a new subscriber was added in `newState` caused by a state change,
assigning `subscriptions = filtered` would then discard that new
subscription.

Since `subscriptions` is a value type, we're safe to mutate in the
`forEach`, so we can fix this bug by never holding a copy of subscriptions,
instead just mutating in-place. If we changed `subscriptions` to a
reference type we would need to enumerate a copy instead.

Performance tests match baselines within allowable margins for error.