Fix regression with `automaticallySkipsRepeats` when selecting Equatable state in Non-Equatable root state
If one's Store's base state did not conform to `Equatable`, the appropriate sub-state selection which skips repeats would not be selected.

This change ensures that `automaticallySkipsRepeats` works as expected in this case.

Fixes #394

Regression was caused in https://github.com/ReSwift/ReSwift/commit/bdb648f1b7817dfe31c12edac0fa2c475643bf66 which was merged in #262