Use GCD instead of deprecated OS level atomic API (#446)
https://github.com/ReSwift/ReSwift/pull/446

Adding concurrency handling was a good call in #341. However, opting for GCD is always better because speed is not always the factor to measure performance. Resource management such as memory and CPU is fine tuned with GCD. For example, `OS` level API's can use 1000+% CPU while GCD will make sure it doesn't go too far over 100%. In the GCD implementation provided here, concurrent reads are happening for fast performance while reading, but a barrier is placed around the writes so no reads or writes can occur during a write. Also `OSAtomicTestAndSet` [seems to be deprecated](https://stackoverflow.com/questions/44685236/swift-3-atomic-boolean).

I added a link in the `Synchronized` utility for reference in case anyone in the future has questions about it or plan to change it, but I can remove it if you like. I've also added unit tests around the synchronization itself, but will a low number to not hold up the unit tests.