Fix bug where automaticallySkipsRepeats is ignored on generic StoreType
If a user was to make a subscription to equatable state using a generic `S: StoreType`, swift would use the default implementation in the store rather than the `SubState: Equatable` extension.

By adding this variation to the `StoreType` protocol, we can ensure that the correct function is called.

A unit test has been added that fails before the protocol addition, and passes after.